requirements.txt
streamlit>=1.44.0
requests>=2.32.0
pandas>=2.2.0
plotly>=5.24.0
python-dotenv>=1.0.1
scikit-learn>=1.5.0
pykrx>=1.2.4
Deprecated>=1.3.1
multipledispatch>=1.0.0
wrapt>=2.1.2
